curupira/backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
gherkin-official==29.0.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
parse==1.20.2
parse_type==0.6.6
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytest-bdd==8.1.0
pytest-django==4.11.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.8.2
PyJWT==2.8.0
